latest mvp

+++ telegram_bots/modules/booking/handlers.py
@@ -9,32 +9,37 @@ from telegram_bots.modules.booking.data_structures import BookingMenu
 from telegram_bots.bots import bot
 
 
-test_user_id = 99999
+# test_user_id = 1
 messages = BookingMessages()
 
 
 async def get_bookings(message: types.Message):
-    bookings = await booking_repository_abstraction.get_users_bookings(test_user_id)
-  
-    user_boking_viewer = booking_viewer_repository.read(message.chat.id)
+    bookings = await booking_repository_abstraction.get_users_bookings(message.chat.id)
+    print('bookings', bookings)
 
-    if not user_boking_viewer:
-        user_boking_viewer = BookingMenu(page_id=0, current_message_id=-1)
-        booking_viewer_repository.create(tg_id=message.chat.id, data=user_boking_viewer)
-        cur_page_id = 0
+    if len(bookings):
+        user_boking_viewer = booking_viewer_repository.read(message.chat.id)
+
+        if not user_boking_viewer:
+            user_boking_viewer = BookingMenu(page_id=0, current_message_id=-1)
+            booking_viewer_repository.create(tg_id=message.chat.id, data=user_boking_viewer)
+            cur_page_id = 0
+        else:
+            await bot.delete_message(message.chat.id, user_boking_viewer.current_message_id)
+            cur_page_id = user_boking_viewer.page_id
+
+        
+
+        booking_message_text, buttons = messages.booking_view_menu_message(bookings, user_boking_viewer.page_id)
+        
+        keyboard = types.InlineKeyboardMarkup(row_width=3)
+        keyboard.add(*buttons)
+        
+        answer_msg = await message.answer(booking_message_text, reply_markup=keyboard)
+        booking_viewer_repository.update(tg_id=message.chat.id, data=BookingMenu(page_id=cur_page_id, current_message_id=answer_msg.message_id))
+
     else:
-        await bot.delete_message(message.chat.id, user_boking_viewer.current_message_id)
-        cur_page_id = user_boking_viewer.page_id
-
-    booking_message_text, buttons = messages.booking_view_menu_message(bookings, user_boking_viewer.page_id)
-    
-    keyboard = types.InlineKeyboardMarkup(row_width=3)
-    keyboard.add(*buttons)
-    
-    answer_msg = await message.answer(booking_message_text, reply_markup=keyboard)
-    booking_viewer_repository.update(tg_id=message.chat.id, data=BookingMenu(page_id=cur_page_id, current_message_id=answer_msg.message_id))
-
-
+        await message.answer('У вас нет актуальных записей')
 
 
 async def cancel_booking_command_handler(message: types.Message):
@@ -57,13 +62,16 @@ async def cancel_booking_command_handler(message: types.Message):
 
 
 async def approve_booking_delete(call: types.CallbackQuery):
+    await booking_repository_abstraction.delete_booking(bookigs_deleter_repository.read(call.message.chat.id))
     await call.message.answer('Запись отменена')
     bookigs_deleter_repository.delete(call.message.chat.id)
-
+    await bot.delete_message(call.message.chat.id, call.message.message_id)
+    
 
 async def deny_booking_delete(call: types.CallbackQuery):
     await call.message.answer('Окей, ничего не отменяем')
     bookigs_deleter_repository.delete(call.message.chat.id)
+    await bot.delete_message(call.message.chat.id, call.message.message_id)
 
 
 def register_booking_handlers(dp: Dispatcher):
